+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Releases
 
 - **Added:** `DOC007` to check for misplaced sections in docstrings.
+- **Added:** Option `--quiet` to suppress output, except statistics if `--statistics` is set.
 
 ### v0.0.5 (2025-09-24)